Add CountTrue by mgriffin
